requirements.txt
geopandas>=1.1.3
ollama>=0.4.0
Pillow>=11.1.0
numpy>=2.1.3
pandas>=2.2.3
plotly>=5.24.1
pydantic>=2.10.3
pytest>=8.3.4
requests>=2.32.3
streamlit>=1.45.1
pyyaml>=6.0
folium>=0.19.0
streamlit-folium>=0.25.0
